requirements.txt
langchain>=0.2.11
langgraph>=0.2.14
langchain-openai>=0.1.22
langsmith>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-chroma>=0.2.0
chromadb>=0.4.0,<0.5.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.1
rich>=13.7.1
typer>=0.12.3
streamlit>=1.32.0
streamlit-chat>=0.1.1
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
